+++ main.tf
@@ -28,7 +28,27 @@ terraform {
 }
 
 locals {
-  cloud                          = "AWS"
-  confluent_cloud_secrets_prefix = "/confluent_cloud_resource"
-  snowflake_secrets_prefix       = "/snowflake_resource"
+  cloud                         = "AWS"
+  secrets_insert                = "apache_flink_kickstarter"
+  confluent_secrets_path_prefix = "/confluent_cloud_resource/${local.secrets_insert}"
+  snowflake_secrets_path_prefix = "/snowflake_resource/${local.secrets_insert}"
+}
+
+
+# Create the Snowflake user RSA keys pairs
+module "snowflake_user_rsa_key_pairs_rotation" {   
+    source  = "github.com/j3-signalroom/iac-snowflake-user-rsa_key_pairs_rotation-tf_module"
+
+    # Required Input(s)
+    aws_region           = var.aws_region
+    aws_account_id       = var.aws_account_id
+    snowflake_account    = jsondecode(data.aws_secretsmanager_secret_version.admin_public_keys.secret_string)["account"]
+    service_account_user = var.service_account_user
+
+    # Optional Input(s)
+    secret_insert             = "/${local.secrets_insert}"
+    day_count                 = var.day_count
+    aws_lambda_memory_size    = var.aws_lambda_memory_size
+    aws_lambda_timeout        = var.aws_lambda_timeout
+    aws_log_retention_in_days = var.aws_log_retention_in_days
 }